+++ src/Config/Symplify/MonorepoBuilder/DataSources/DataToAppendAndRemoveConfig.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PoP\PoP\Config\Symplify\MonorepoBuilder;
+namespace PoP\PoP\Config\Symplify\MonorepoBuilder\DataSources;
 
 class DataToAppendAndRemoveConfig
 {
